removed duplicate code.
